+++ CHANGELOG.md
@@ -116,6 +116,7 @@
 - Fix the paths of old pictures of foodsharepoints #1490 !2519 @alex.simm
 - Prevent invalid states in notification icons #1465 !2536 @AntonBallmaier
 - Prevent errors in the calendar export caused by events with wrong dates #1301 !2532 @alex.simm
+- Sanitize public info in store to prevent xss !2538 @chriswalg
 
 ## Refactoring
 - Currently running events are shown within the event overview until their end !2162 @merhoc